transfer copy sidepanel into its own component
